+++ code/engine.vc2008/xrEngine/Environment.h
@@ -75,7 +75,7 @@ public:
 		typedef xr_vector<ref_sound>	sounds_type;
 
 		void					load					(CInifile& config, LPCSTR sect);
-		ref_sound&				get_rnd_sound			()	{return sounds()[Random.randI(sounds().size())];}
+		ref_sound&				get_rnd_sound			()	{return sounds()[Random.randI(s32(sounds().size()))];}
 		u32						get_rnd_sound_time		()	{return (m_sound_period.z < m_sound_period.w) ? Random.randI(m_sound_period.z,m_sound_period.w) : 0;}
 		u32						get_rnd_sound_first_time()	{return (m_sound_period.x < m_sound_period.y) ? Random.randI(m_sound_period.x,m_sound_period.y) : 0;}
 		float					get_rnd_sound_dist		()	{return (m_sound_dist.x < m_sound_dist.y) ? Random.randF(m_sound_dist.x, m_sound_dist.y) : 0;}
@@ -99,13 +99,8 @@ public:
 	IC const shared_str&	name				()	{return m_load_section;}
 	IC const shared_str&	get_ambients_config_filename ()	{return m_ambients_config_filename;}
 
-	INGAME_EDITOR_VIRTUAL	void	load		(
-								CInifile& ambients_config,
-								CInifile& sound_channels_config,
-								CInifile& effects_config,
-								const shared_str& section
-							);
-	IC SEffect*				get_rnd_effect		()	{return effects().empty()?0:effects()[Random.randI(effects().size())];}
+	INGAME_EDITOR_VIRTUAL	void load(CInifile& ambients_config, CInifile& sound_channels_config, CInifile& effects_config, const shared_str& section);
+	IC SEffect*				get_rnd_effect		()	{return effects().empty()?0:effects()[Random.randI(s32(effects().size()))];}
 	IC u32					get_rnd_effect_time ()	{return Random.randI(m_effect_period.x, m_effect_period.y);}
 
 	INGAME_EDITOR_VIRTUAL	SEffect*		create_effect			(CInifile& config, LPCSTR id);
